spec/support: extract OctokitFetcher to its own file (TODO 13)
The fixture_downloader.rb file had grown to ~385 lines with two distinct
concerns bundled together:

  1. Retry / routing loop (the Downloader class) - backoff, Retry-After,
     permanent-vs-transient classification.
  2. Octokit URL dispatch (the OctokitFetcher module) - release-asset,
     raw-file, contents-API routing.

These are MECE-distinct. Bundling them muddied the boundary; future
changes to Octokit routing shouldn't require reading the retry loop.

Extracted OctokitFetcher to spec/support/fixture_downloader/octokit_fetcher.rb.
The main file is now 296 lines focused on the Downloader class; the
fetcher is 89 lines focused on URL-pattern dispatch.

Specs: 21 examples, all pass unchanged.
EOF
2>&1 | tail -3
